Update Nginx, PostgreSQL and RabbitMQ role location and/or version from the Ansible Requirements file.
As Ansible 2 is the main version of Ansible, some roles need to be
updated following the new version recommendation.
For the RabbitMQ role, some people manage to fix the errors, but the
fix has not been merged upstream, so we forked the repository and made
our own changes.
Others Ansible roles have been updated.

Close: quarkslab/irma-ansible#129