Move the IRMA Web-UI to its own directory.
As the Web-ui is now in it's own directory, there is a need to update
the directory layout for Ansible.
Moreover, we changed how ansible manage allinone_{dev.prod}
and {dev,prod} environments, to include the fact that the web-ui can be
install as another nginx site. In this particular case, there is a need
to activate CORS HTTP Headers.