Fix configuration for ReadTheDocs build.
We were using an old version of the sphinx_rtd_theme which break the
design on ReadTheDocs website.

This commit remove the old template and modify the config file, which is
used by ReadTheDocs to generate the online documentation.
If an user want to locally generate the documentation, he will need to
install the dependencies (see requirements.txt) before running `make
html`.